resolved queue thumbnails
