--- AmahiAnywhere/AmahiAnywhere/Cells/AudioTrackTableViewCell.xib
@@ -1,79 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="16096" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina4_7" orientation="portrait" appearance="light"/>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="16087"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="AudioTrackTableViewCell" customModule="AmahiAnywhere" customModuleProvider="target"/>
-        <view contentMode="scaleToFill" id="GZL-Jr-xio" customClass="AudioTrackTableViewCell" customModule="AmahiAnywhere" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
-            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-            <subviews>
-                <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="jYn-zZ-hLG">
-                    <rect key="frame" x="40" y="0.0" width="335" height="667"/>
-                    <subviews>
-                        <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleToFill" horizontalHuggingPriority="251" verticalHuggingPriority="240" image="musicPlayerArtWork" translatesAutoresizingMaskIntoConstraints="NO" id="PHF-So-7Ql">
-                            <rect key="frame" x="5" y="5" width="50.5" height="657"/>
-                        </imageView>
-                        <stackView opaque="NO" contentMode="scaleToFill" axis="vertical" distribution="fillEqually" spacing="5" translatesAutoresizingMaskIntoConstraints="NO" id="Lvw-uD-Nsr">
-                            <rect key="frame" x="60.5" y="5" width="269.5" height="657"/>
-                            <subviews>
-                                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Title" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="fQr-Vb-U4j">
-                                    <rect key="frame" x="0.0" y="0.0" width="269.5" height="326"/>
-                                    <fontDescription key="fontDescription" type="boldSystem" pointSize="14"/>
-                                    <nil key="textColor"/>
-                                    <nil key="highlightedColor"/>
-                                    <variation key="heightClass=regular-widthClass=regular">
-                                        <fontDescription key="fontDescription" type="boldSystem" pointSize="18"/>
-                                    </variation>
-                                </label>
-                                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Artist" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="ayg-Db-VQS">
-                                    <rect key="frame" x="0.0" y="331" width="269.5" height="326"/>
-                                    <fontDescription key="fontDescription" type="system" pointSize="12"/>
-                                    <color key="textColor" systemColor="secondaryLabelColor" red="0.23529411759999999" green="0.23529411759999999" blue="0.26274509800000001" alpha="0.59999999999999998" colorSpace="custom" customColorSpace="sRGB"/>
-                                    <nil key="highlightedColor"/>
-                                    <variation key="heightClass=regular-widthClass=regular">
-                                        <fontDescription key="fontDescription" type="system" pointSize="16"/>
-                                    </variation>
-                                </label>
-                            </subviews>
-                        </stackView>
-                    </subviews>
-                    <constraints>
-                        <constraint firstAttribute="bottom" secondItem="PHF-So-7Ql" secondAttribute="bottom" constant="5" id="Qfo-qR-H1A"/>
-                        <constraint firstItem="Lvw-uD-Nsr" firstAttribute="top" secondItem="jYn-zZ-hLG" secondAttribute="top" constant="5" id="Z8g-Db-qy6"/>
-                        <constraint firstAttribute="trailing" secondItem="Lvw-uD-Nsr" secondAttribute="trailing" constant="5" id="cUG-b1-0d8"/>
-                        <constraint firstItem="PHF-So-7Ql" firstAttribute="width" secondItem="jYn-zZ-hLG" secondAttribute="width" multiplier="0.15" id="dMX-AV-gRC"/>
-                        <constraint firstItem="Lvw-uD-Nsr" firstAttribute="leading" secondItem="PHF-So-7Ql" secondAttribute="trailing" constant="5" id="fXo-2G-X2s"/>
-                        <constraint firstItem="PHF-So-7Ql" firstAttribute="leading" secondItem="jYn-zZ-hLG" secondAttribute="leading" constant="5" id="gwG-xE-AhV"/>
-                        <constraint firstItem="PHF-So-7Ql" firstAttribute="top" secondItem="jYn-zZ-hLG" secondAttribute="top" constant="5" id="pmE-n6-RkT"/>
-                        <constraint firstAttribute="bottom" secondItem="Lvw-uD-Nsr" secondAttribute="bottom" constant="5" id="r2Y-2c-DYh"/>
-                    </constraints>
-                </view>
-            </subviews>
-            <color key="backgroundColor" systemColor="systemBackgroundColor" cocoaTouchSystemColor="whiteColor"/>
-            <constraints>
-                <constraint firstItem="jYn-zZ-hLG" firstAttribute="top" secondItem="KAH-n4-JwX" secondAttribute="top" id="QJJ-GD-PF8"/>
-                <constraint firstItem="KAH-n4-JwX" firstAttribute="bottom" secondItem="jYn-zZ-hLG" secondAttribute="bottom" id="gsp-gQ-fTI"/>
-                <constraint firstItem="jYn-zZ-hLG" firstAttribute="leading" secondItem="KAH-n4-JwX" secondAttribute="leading" constant="40" id="jDA-rQ-0Qf"/>
-                <constraint firstItem="KAH-n4-JwX" firstAttribute="trailing" secondItem="jYn-zZ-hLG" secondAttribute="trailing" id="mC4-B1-mXD"/>
-            </constraints>
-            <viewLayoutGuide key="safeArea" id="KAH-n4-JwX"/>
-            <connections>
-                <outlet property="artistLabel" destination="ayg-Db-VQS" id="29I-fN-B7y"/>
-                <outlet property="containerView" destination="jYn-zZ-hLG" id="VhZ-gI-Ayn"/>
-                <outlet property="thumbnailView" destination="PHF-So-7Ql" id="per-V7-ZOy"/>
-                <outlet property="titleLabel" destination="fQr-Vb-U4j" id="VqX-IW-BUj"/>
-            </connections>
-            <point key="canvasLocation" x="1285.5999999999999" y="1079.9100449775112"/>
-        </view>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-    </objects>
-    <resources>
-        <image name="musicPlayerArtWork" width="300" height="300"/>
-    </resources>
-</document>
