+++ AmahiAnywhere/AmahiAnywhere/Cells/QueueItemTableViewCell.xib
@@ -1,0 +1,81 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="16096" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="retina6_1" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="16087"/>
+        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <tableViewCell contentMode="scaleToFill" selectionStyle="default" indentationWidth="10" id="KGk-i7-Jjw" customClass="QueueItemTableViewCell" customModule="AmahiAnywhere" customModuleProvider="target">
+            <rect key="frame" x="0.0" y="0.0" width="320" height="44"/>
+            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+            <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="KGk-i7-Jjw" id="H2p-sc-9uM">
+                <rect key="frame" x="0.0" y="0.0" width="320" height="44"/>
+                <autoresizingMask key="autoresizingMask"/>
+                <subviews>
+                    <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="XQD-ho-HlE">
+                        <rect key="frame" x="0.0" y="0.0" width="320" height="44"/>
+                        <subviews>
+                            <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleToFill" horizontalHuggingPriority="251" verticalHuggingPriority="240" image="musicPlayerArtWork" translatesAutoresizingMaskIntoConstraints="NO" id="BMk-mY-QLE">
+                                <rect key="frame" x="5" y="5" width="48" height="34"/>
+                            </imageView>
+                            <stackView opaque="NO" contentMode="scaleToFill" axis="vertical" distribution="fillEqually" spacing="5" translatesAutoresizingMaskIntoConstraints="NO" id="hft-1p-iJZ">
+                                <rect key="frame" x="58" y="5" width="257" height="34"/>
+                                <subviews>
+                                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Title" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="Brt-5Z-69O">
+                                        <rect key="frame" x="0.0" y="0.0" width="257" height="14.5"/>
+                                        <fontDescription key="fontDescription" type="boldSystem" pointSize="14"/>
+                                        <nil key="textColor"/>
+                                        <nil key="highlightedColor"/>
+                                        <variation key="heightClass=regular-widthClass=regular">
+                                            <fontDescription key="fontDescription" type="boldSystem" pointSize="18"/>
+                                        </variation>
+                                    </label>
+                                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Artist" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="Axz-Jg-xj3">
+                                        <rect key="frame" x="0.0" y="19.5" width="257" height="14.5"/>
+                                        <fontDescription key="fontDescription" type="system" pointSize="12"/>
+                                        <color key="textColor" systemColor="secondaryLabelColor" red="0.23529411759999999" green="0.23529411759999999" blue="0.26274509800000001" alpha="0.59999999999999998" colorSpace="custom" customColorSpace="sRGB"/>
+                                        <nil key="highlightedColor"/>
+                                        <variation key="heightClass=regular-widthClass=regular">
+                                            <fontDescription key="fontDescription" type="system" pointSize="16"/>
+                                        </variation>
+                                    </label>
+                                </subviews>
+                            </stackView>
+                        </subviews>
+                        <constraints>
+                            <constraint firstAttribute="trailing" secondItem="hft-1p-iJZ" secondAttribute="trailing" constant="5" id="HF0-DA-0bi"/>
+                            <constraint firstAttribute="bottom" secondItem="hft-1p-iJZ" secondAttribute="bottom" constant="5" id="JkX-cq-vR1"/>
+                            <constraint firstAttribute="bottom" secondItem="BMk-mY-QLE" secondAttribute="bottom" constant="5" id="Z8b-9a-wHT"/>
+                            <constraint firstItem="hft-1p-iJZ" firstAttribute="leading" secondItem="BMk-mY-QLE" secondAttribute="trailing" constant="5" id="aIG-UE-FuC"/>
+                            <constraint firstItem="hft-1p-iJZ" firstAttribute="top" secondItem="XQD-ho-HlE" secondAttribute="top" constant="5" id="bmy-to-jYo"/>
+                            <constraint firstItem="BMk-mY-QLE" firstAttribute="leading" secondItem="XQD-ho-HlE" secondAttribute="leading" constant="5" id="ejm-XT-RGN"/>
+                            <constraint firstItem="BMk-mY-QLE" firstAttribute="width" secondItem="XQD-ho-HlE" secondAttribute="width" multiplier="0.15" id="gt9-49-bJX"/>
+                            <constraint firstItem="BMk-mY-QLE" firstAttribute="top" secondItem="XQD-ho-HlE" secondAttribute="top" constant="5" id="wGR-B6-Hf8"/>
+                        </constraints>
+                    </view>
+                </subviews>
+                <constraints>
+                    <constraint firstItem="XQD-ho-HlE" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="top" id="A0m-rE-suT"/>
+                    <constraint firstAttribute="trailing" secondItem="XQD-ho-HlE" secondAttribute="trailing" id="sY4-gs-ZD0"/>
+                    <constraint firstAttribute="bottom" secondItem="XQD-ho-HlE" secondAttribute="bottom" id="tk7-R3-fWg"/>
+                    <constraint firstItem="XQD-ho-HlE" firstAttribute="leading" secondItem="H2p-sc-9uM" secondAttribute="leading" id="yGl-4L-T5e"/>
+                </constraints>
+            </tableViewCellContentView>
+            <viewLayoutGuide key="safeArea" id="njF-e1-oar"/>
+            <connections>
+                <outlet property="artistLabel" destination="Axz-Jg-xj3" id="j9w-UO-fzg"/>
+                <outlet property="thumbnailView" destination="BMk-mY-QLE" id="20e-5V-yQL"/>
+                <outlet property="titleLabel" destination="Brt-5Z-69O" id="lW1-Ph-jZF"/>
+            </connections>
+            <point key="canvasLocation" x="139" y="113"/>
+        </tableViewCell>
+    </objects>
+    <resources>
+        <image name="musicPlayerArtWork" width="300" height="300"/>
+    </resources>
+</document>
